add main controller

+++ src/main/java/com/unproject/onlineMagazine/model/dao/Product.java
@@ -11,9 +11,9 @@ public class Product {
     private Long id;
     private String product_type;
     private String description;
-    private int weight;
+    private double weight;
 
-    public Product(String product_type, String description, int weight) {
+    public Product(String product_type, String description, double weight) {
         this.product_type = product_type;
         this.description = description;
         this.weight = weight;
